Add xdebug support for vim.
The recommended location for the xdebug scripts is plugin in .vim, but
I'd rather be more explicit than that since I'm using Pathogen to manage
my plugins. So, the files are in xdebugger, and debugger.vim has been
modified from the original to point there. I also have <Leader>x mapped
to load the plugin so it's not loaded every time I run vim.